Updated Pipfile, Pipfile.lock to lock in rx v1.x
The current codebase was designed with the rx v1.x model. Updating to
v3.x would break many things. When a new package was being installed
using pipenv, it would update rx to 3.x as well during the process of
creating Pipfile.lock. So it's necessary to pin down rx < 3.0